Adds ACRE racks and intercoms to ACH-47 "Guns-aGo-Go", OH-6A and CH-34 Seahorse
