--- addons/sogpf_acre/inc_CfgVehicles_Heli.hpp
@@ -1,332 +1,0 @@
-class vn_helicopter_base;
-
-// AH-1 Variants
-class vn_air_ah1g_01_base: vn_helicopter_base {
-    acre_hasInfantryPhone = 0;
-
-    class AcreRacks {
-        class GVAR(rack_ah1_fm_comms) {
-            displayName = "FM Communication";
-            shortName = "FM";
-
-            componentName = "ACRE_VRC64";
-            mountedRadio = "ACRE_PRC77";
-            isRadioRemovable = 0;
-            defaultComponents[] = {};
-
-            allowedPositions[] = {
-                "driver"
-            };
-            disabledPositions[] = {};
-
-            intercom[] = {QGVAR(intercom_ah1)};
-        };
-    };
-
-    class AcreIntercoms {
-        class GVAR(intercom_ah1) {
-            displayName = "$STR_ACRE_sys_intercom_crewIntercom";
-            shortName = "$STR_ACRE_sys_intercom_shortCrewIntercom";
-
-            allowedPositions[] = {
-                "crew"
-            };
-            disabledPositions[] = {};
-            masterPositions[] = {
-                {"driver"}
-            };
-
-            limitedPositions[] = {};
-            numLimitedPositions = 0;
-
-            connectedByDefault = 1;
-        };
-    };
-};
-
-// CH-47 Variants
-class vn_air_ch47_base: vn_helicopter_base {
-    acre_hasInfantryPhone = 0;
-
-    class AcreRacks {
-        class GVAR(rack_ch47_fm_comms) {
-            displayName = "FM Communication";
-            shortName = "FM";
-
-            componentName = "ACRE_VRC64";
-            mountedRadio = "ACRE_PRC77";
-            isRadioRemovable = 0;
-            defaultComponents[] = {};
-
-            allowedPositions[] = {
-                "driver",
-                {"turret", {0}}
-            };
-            disabledPositions[] = {};
-
-            intercom[] = {QGVAR(intercom_ch47)};
-        };
-    };
-
-    class AcreIntercoms {
-        class GVAR(intercom_ch47) {
-            displayName = "$STR_ACRE_sys_intercom_crewIntercom";
-            shortName = "$STR_ACRE_sys_intercom_shortCrewIntercom";
-
-            allowedPositions[] = {
-                "crew"
-            };
-            disabledPositions[] = {};
-            masterPositions[] = {
-                {"gunner", {3}}
-            };
-
-            limitedPositions[] = {
-                {"cargo", "all"}, 
-                {"ffv", "all"}
-            };
-            numLimitedPositions = 4;
-
-            connectedByDefault = 1;
-        };
-    };
-};
-
-// UH-1 Variants
-class vn_air_uh1_01_base;
-class vn_air_uh1d_base;
-
-class vn_air_uh1c_base: vn_air_uh1_01_base {
-    acre_hasInfantryPhone = 0;
-
-    class AcreRacks {
-        class GVAR(rack_uh1_fm_comms) {
-            displayName = "FM Communication";
-            shortName = "FM";
-
-            componentName = "ACRE_VRC64";
-            mountedRadio = "ACRE_PRC77";
-            isRadioRemovable = 0;
-            defaultComponents[] = {};
-
-            allowedPositions[] = {
-                "driver", 
-                {"gunner", {0}}
-            };
-            disabledPositions[] = {};
-
-            intercom[] = {QGVAR(intercom_uh1)};
-        };
-    };
-
-    class AcreIntercoms {
-        class GVAR(intercom_uh1) {
-            displayName = "$STR_ACRE_sys_intercom_crewIntercom";
-            shortName = "$STR_ACRE_sys_intercom_shortCrewIntercom";
-
-            allowedPositions[] = {
-                "crew"
-            };
-            disabledPositions[] = {};
-            masterPositions[] = {
-                {"gunner", {0}}
-            };
-
-            limitedPositions[] = {
-                {"cargo", "all"}, 
-                {"ffv", "all"}
-            };
-            numLimitedPositions = 1;
-
-            connectedByDefault = 1;
-        };
-    };
-};
-
-class vn_air_uh1c_doorguns_base: vn_air_uh1c_base {
-    acre_hasInfantryPhone = 0;
-
-    class AcreRacks {
-        class GVAR(rack_uh1_fm_comms) {
-            displayName = "FM Communications";
-            shortName = "FM";
-
-            componentName = "ACRE_VRC64";
-            mountedRadio = "ACRE_PRC77";
-            isRadioRemovable = 0;
-            defaultComponents[] = {};
-
-            allowedPositions[] = {
-                "driver", 
-                {"gunner", {0}}
-            };
-            disabledPositions[] = {};
-
-            intercom[] = {QGVAR(intercom_uh1)};
-        };
-    };
-
-    class AcreIntercoms {
-        class GVAR(intercom_uh1) {
-            displayName = "$STR_ACRE_sys_intercom_crewIntercom";
-            shortName = "$STR_ACRE_sys_intercom_shortCrewIntercom";
-
-            allowedPositions[] = {
-                "crew"
-            };
-            disabledPositions[] = {};
-            masterPositions[] = {
-                {"turret", {1}}
-            };
-
-            limitedPositions[] = {
-                {"cargo", "all"}, 
-                {"ffv", "all"}
-            };
-            numLimitedPositions = 2;
-
-            connectedByDefault = 1;
-        };
-    };
-};
-
-class vn_air_uh1d_01_base: vn_air_uh1d_base {
-    acre_hasInfantryPhone = 0;
-
-    class AcreRacks {
-        class GVAR(rack_uh1_fm_comms) {
-            displayName = "FM Communication";
-            shortName = "FM";
-
-            componentName = "ACRE_VRC64";
-            mountedRadio = "ACRE_PRC77";
-            isRadioRemovable = 0;
-            defaultComponents[] = {};
-
-            allowedPositions[] = {
-                "driver", 
-                {"gunner", {0}}
-            };
-            disabledPositions[] = {};
-
-            intercom[] = {QGVAR(intercom_uh1)};
-        };
-    };
-
-    class AcreIntercoms {
-        class GVAR(intercom_uh1) {
-            displayName = "$STR_ACRE_sys_intercom_crewIntercom";
-            shortName = "$STR_ACRE_sys_intercom_shortCrewIntercom";
-
-            allowedPositions[] = {
-                "crew"
-            };
-            disabledPositions[] = {};
-            masterPositions[] = {
-                {"gunner", {0}}
-            };
-
-            limitedPositions[] = {
-                {"cargo", "all"}, 
-                {"ffv", "all"}
-            };
-            numLimitedPositions = 1;
-
-            connectedByDefault = 1;
-        };
-    };
-};
-
-class vn_air_uh1d_02_base: vn_air_uh1d_base {
-    acre_hasInfantryPhone = 0;
-
-    class AcreRacks {
-        class GVAR(rack_uh1_fm_comms) {
-            displayName = "FM Communication";
-            shortName = "FM";
-
-            componentName = "ACRE_VRC64";
-            mountedRadio = "ACRE_PRC77";
-            isRadioRemovable = 0;
-            defaultComponents[] = {};
-
-            allowedPositions[] = {
-                "driver", 
-                {"gunner", {0}}
-            };
-            disabledPositions[] = {};
-
-            intercom[] = {QGVAR(intercom_uh1)};
-        };
-    };
-
-    class AcreIntercoms {
-        class GVAR(intercom_uh1) {
-            displayName = "$STR_ACRE_sys_intercom_crewIntercom";
-            shortName = "$STR_ACRE_sys_intercom_shortCrewIntercom";
-
-            allowedPositions[] = {
-                "crew"
-            };
-            disabledPositions[] = {};
-            masterPositions[] = {
-                {"turret", {1}}
-            };
-
-            limitedPositions[] = {
-                {"cargo", "all"}, 
-                {"ffv", "all"}
-            };
-            numLimitedPositions = 1;
-
-            connectedByDefault = 1;
-        };
-    };
-};
-
-class vn_air_uh1d_04_base: vn_air_uh1d_base {
-    acre_hasInfantryPhone = 0;
-
-    class AcreRacks {
-        class GVAR(rack_uh1_fm_comms) {
-            displayName = "FM Communication";
-            shortName = "FM";
-
-            componentName = "ACRE_VRC64";
-            mountedRadio = "ACRE_PRC77";
-            isRadioRemovable = 0;
-            defaultComponents[] = {};
-
-            allowedPositions[] = {
-                "driver", 
-                {"gunner", {0}}
-            };
-            disabledPositions[] = {};
-
-            intercom[] = {QGVAR(intercom_uh1)};
-        };
-    };
-
-    class AcreIntercoms {
-        class GVAR(intercom_uh1) {
-            displayName = "$STR_ACRE_sys_intercom_crewIntercom";
-            shortName = "$STR_ACRE_sys_intercom_shortCrewIntercom";
-
-            allowedPositions[] = {
-                "crew"
-            };
-            disabledPositions[] = {};
-            masterPositions[] = {
-                {"gunner", {0}}
-            };
-
-            limitedPositions[] = {
-                {"cargo", "all"}, 
-                {"ffv", "all"}
-            };
-            numLimitedPositions = 1;
-
-            connectedByDefault = 1;
-        };
-    };
-};
